Compare two variant call sets

This function takes two samples as input and compares them using dplyr::*_join() functions. The results of the joins are captured and returned in a list. This is an internal function.

Usage

# S3 method for hybcap
compare_vars(
  mdlvalr_list,
  pipeline = "hybcap",
  match_by_colnames = c("CHROM", "POS", "REF", "ALT", "GENE", "EXON"),
  suffix_names = c(".sample1", ".sample2")
)

Arguments

match_by_colnames

Vector of column names to use for matching. Passed to by = parameter in join functions. default: c("CHROM", "POS", "REF", "ALT", "GENE", "EXON").

suffix_names

Suffix names added to column names after running join functions. Vector of length two, default: c(".sample_1", ".sample_2").

sample_1_name

String for sample_1 name.

sample_1_var

Tibble of sample_1 variants.

sample_1_cov

Tibble of sample_1 exon coverage.

sample_2_name

String for sample_2 name.

sample_2_var

Tibble of sample_2 variants.

sample_2_cov

Tibble of sample_2 exon coverage.

min_vaf

Variant tables are filtered based on minimum variant allele frequency (VAF). Variants with VAF less than or equal to this value are not used in comparison. default: 0.05.

min_fraction_125x

Exon tables are filtered based on the minimum fraction of bases covered at 125x. default: 0.9.

Value

Returns a named list of tibbles or integers tibble sizes.