Filter compared data into sub-tables based on vairant labels

Filter compared data into sub-tables based on vairant labels

Usage

filter_data(mdlvalr_list, pipeline = NULL)

# S3 method for hybcap
filter_data(mdlvalr_list, pipeline = "hybcap")

Arguments

mdlvalr_list

A standard mdlvalr_list object (as generated by add_var_labels()). The class of this object can be used to automatically determine which function methods will be used.

pipeline

Standard name of the pipeline from which the data were derived.

Value

Returns a standard mdlvalr_list object (i.e. a regular R list of lists) that contains tables as named list elements. The class of the retuned object is updated to include the name of the pipeline. This facilitates use of alternative methods in downstream functions.

Methods

This function is a generic, which means that different implementations (methods) can be executed for different classes. The class of the first argument will determine which method is used. See the documentation of individual methods for extra arguments and differences in behaviour.

Examples

class(mdlvalr_list)
#> Error in eval(expr, envir, enclos): object 'mdlvalr_list' not found
# [1] "hybcap" "list"
{
mdlvalr_list <- filter_data(
    mdlvalr_list,
    pipeline = "hybcap")
}
#> Error in filter_data(mdlvalr_list, pipeline = "hybcap"): object 'mdlvalr_list' not found