Validate an input table of sample info and generate sample sheet

validate_samples() takes an input table of sample data and uses important columns to generate a valid sample sheet that can be used as input to get_data().

Usage

validate_samples(
  samples_tbl,
  sample_name,
  sample_group,
  comparison_group,
  var_path,
  cov_path,
  pipeline
)

Arguments

samples_tbl

Tibble of sample info. Requires five columns (listed below).

sample_name

Column name that specifies the sample names.

sample_group

Column name that specifies the sample_group.

comparison_group

Column name that specifies the comparison_group.

var_path

Column name that specifies the variants table file path.

cov_path

Column name that specifies the coverage table file path.

pipeline

Name of pipeline that produced the data. One of c("hybcap", "exome", "germline").

Value

Returns a sample sheet (tibble) of comparisons to make. The class of the retuned object is updated to include the name of the pipeline. This facilitates use of alternative methods in downstream functions. This can be used as input to get_data().

Examples

# The package provides example data `samples_tbl` that can be used for demo.
# Call the example data to make it available. 
invisible(samples_tbl)
sample_sheet <- validate_samples(samples_tbl,
   sample_name = "sample_name",
   sample_group = "sample_name_short",
   comparison_group = "group",
   var_path = "filtered_path",
   cov_path = "coverage_path",
   pipeline = "hybcap")
#> Error in samples_tbl %>% dplyr::select(all_of(c(sample_name, sample_group,     comparison_group, var_path, cov_path))) %>% magrittr::set_colnames(c("sample_name",     "sample_group", "comparison_group", "var_path", "cov_path")) %>%     dplyr::mutate(comparison_name = glue("{sample_group}_{comparison_group}")): could not find function "%>%"